Finished Visualizations for Squads
Haven't tested with defending squads on the board but it should work. The squads of each force are the exact same just with one variable either a zero or a one so that should either be easy to fix if it is broken or already working. I will check that tomorrow. the models have their feet and tracks in the ground, so I will raise those models a little bit later, but that again is easy.

The only other thing to change is fighting animations, but that is not really necessary now and even when I do add it, it will only be a single day of work I would imagine. most of that functionality already exists. I would just need to extend the tank and aircraft animations script and then figure out how to sqwap animations for the soldiers. But that should be at max a days worth of work. But again not important right now.