Fixed Visualizations for Defenders
Changed the visualization up quite a bit. But the defender squads are viewed properly now. Ended up having to change quite a bit because it was honestly a bit messy the way I was adding defender squads in along with the others. So now deployed and defending squads are in a Dict<string, int[]> and the clientInterface tracks how many ownedDefendersOnIsland. When it submits to the chain it just doesn't count the first couple of squads in queuedActions.dfnd.sqd via ownedDefendersOnIsland and makes a new list that ommits those.