Tutorial Part 1
Got a lot of the Tutorial done. The scripts and some of the gui stuff. I think tomorrow it should either be completely finished or only some small code bits needed. As it is right now I don't think it will work if you try to boot it. I forgot to check before closing unity.

But the script for tutorials is pretty simple. There is a master tutorial which extends interaction.cs solely to use some of those variables as triggers and there are tutorial prompters which you edit similarily to world buttons. Just put the name of the tutorial stage and then add a bunch of subtutorial prefabs with the button calling the NextTutorial() function of the TutorialPrompter script.

Tutorial.cs triggers tutorialPrompters with InitiateTutorial() and TutorialPrompter.cs iterates through messages in order via NextTutorial(). Tutorial knows which stage we are in by custom conditions in Tutorial but they know which prompter to trigger by the "tutorialName" variable in the prompter variables.

I still have more tutorials for unit purchasing,  resource pools, warbucks, and attacking and defending. But that will probably be done tomorrow, and if not it will be some of the custom trigges conditions that need finishing. The messages are mostly just boring and tedious. Nothing complicated. Literally just texts and some foresight on what the trigger will be.