More Tutorial Stuff
I didn't get all of the tutorials done but I got most of them. All that is left is the ResourcePool tutorials. Also I forgot to add the button operations, so I will need to do that for all of the tutorials I did today, but I might has well do that after I am done with the ResourcePool tutorials.

That should all be done pretty quickly tomorrow, then I need to get the conditions for them programmed out and tested, I don't know how long that will take but I don't think too long. The testing is what will take the longest.