Finished Tutorial All Parts
Finished the tutorial conditions and all of the tiny edits that needed to be made. There were some bugs I corrected as well.

Next step is moving it to the mainnet and fixing the bug that crash the game on close. I'm near 100% sure it is the Stop command from the xaya library. I have to figure out why it is not going through. I may need to update the libXayaGameI am using as it is currently the one from the tutorial which im pretty sure is broken.

But regardless, I think the problem lies in that function and once that is working the sqlite and/or lmdb will work flawlessly as well. I haven't tested reorgs, but I'm just saving the last state so there is no real reason to test the reorgs extensively. Not until we start making more sophisticated undo data.

So other than some sounds and music which joshua is working on, splash screens, and a couple of graphical tweaks, and the above mentioned crashing problem, this is actually pretty much done after I get it on main net.