Settings Complete
All of the settings options are complete. Getting a stack overflow while rounding the FPS slider, but it is not really needed right now, so I have just taken it out instead of trying to track down the bug.

The one bug I do need to track down is the libxayagame bug. I will likely have to recompile the library based off of the latest and I am hoping that doesnt break my shit but yeah, the bug is somewhere in there with the xaaGameService.Stop() function. But once that works, I'm fairly comnfident the sqlite problem will be solved as well.