Changed Constants & Cleaned Folders
I changed the constants from a static class to a normal public class that State holds. I wasnt think and  the other day I realized that the game state processor will only update constants correctly if it is processed from the beginning every time and not when it is loaded from sqlite or lmbd. So I converted everything to reference and accept constants as variable in State

I also cleaned out the staging folder and moved everything to their correct places. Haven't deleted unused files and materials. But I can do that later.