Added Resource Pack Purchase
I have added the functionality of purchasing resource packs for chi. I haven't tested it, but that is pretty simple. I can do that tomorrow. The guys at xaya likely won't be working over the weekend so I will not be able to resolve the SQLite crashing problem. So I will just work on that over the weekend.

A thought just occured to me, I do recall getting the sqlite to work somehow in the mover demo, I'm curious if it was because it was using the electron wallet on the mainnet. I should probably try that out. Maybe it has just been the xayad version I have been using, or quite possibly mayber I'm remembering gettings something else working. It was a while ago.