Fixed Warbux Purchasing for Chi
The warbux purchasing wasn't working initially because I was sending the "sendCoin" json as a string like the normal name_updates but domob informed me that those options need to be sent as a JsonObject unlike the name updates, so that was an easy fix. Since I have time and can't really fix the crashing problem without the assistance of the Xaya Team, I'll be working on the player trading market next.