Completed backend player trading.
All of the backend code for player trading is completed, no testing on it done yet, it will be easier to test with an interface so I will program the front end for it tomorrow and complete the guis for both player trading and warbux purchasing. If there is no problems with the backend, they will both likely be done completely tomorrow.