Redid the save system.
I was using the PlayerPrefs as a save system, but that got less and less useful as I started expanding the functionality and I ended up doing some weird hinky shit to just keep it working. I redid all of it and now it is saving a binary file in LocalLow/MudHeroStudios as preferences.dat. It's pretty much just a json string in a file, but I figure I can use that save system later in different games so it accepts any string and saves it in a file. I'll expand upon it later but for now it works.

I haven't tested a new user name for tutorials though. Only an already existing one. I can do that tomorrow.