Added Remove Squad from island option.
I have added an option to remove squads from previously submitted battle plans. Before there was no way to remove already existing defense squads. It is a small red x over the markers.

Being that Xaya still hasn't figured out the crashing problems with saving, I'll just continue to add stuff until I either don't have anything else to add or they fix it. So I'll be working on a notification system. Right now there is no notification system for when something happens on a block update. So that will be nice for players to know when they have found an island or were rewarded portions of a pool or whatever.