Changed Most Icons to Graphics
Just went through and changed a bunch of the button icons to pictures instead. Also fixed one unit purchasing bug. I had commented out a function that was vital for the "on change" functionality in the input field. So it defaulted to one of the overloads and kept setting everything to type 0 which was riflemen. Easy fix. Just had to uncomment.