Fixed Many Gamestate bugs
There were a couple of game state bugs, one of the was breaking all the versions so I had to correct it immediately.

The market order bug was beacause I was setting the seller with player data.
The resource and warbux pools were not being rewarded because there was no loop for them in the callbacks.
There was a warbux pool bug as well where the depleted contributions were not being cleared.
I changed uPerc in admin commands from int to float, i dont know why it was int but that wouldnt have worked with the percent.