Fixed Squad Send
There was a small error in the client where the command for defense plan was being created and it malformed the command and caused the cancel to not work when you click it if you didnt have enough troops and cloned your squad counts of the first squad when you did have enough troops.

Also this version has form squad tutorial disabled because it was giving errors and I didnt want to fix it just yet or have it messing with the results of my probing so i disabled it temporarily until i can get around to fixing it.