Fixed Some Errors in Attack Tile Menu
There was an error that would delete the island because the attacakble was null and I was only checking for "". There was another string building error where is would say "Island No Attackable Island". getting off for today. The next bug on my list will require refactoring and that will take a bit and I dont want to stop half way through.