Refactored and Fixed Battle Planning.
There were some bugs with units on the battle planning for both attacka nd defend if there were already units on it. They seem to be resolved. I ended up just refactoring it a little bit because it was a bit messy and confusing so I said fuck it, and started the section responsible for the bug from scratch.