Updated StateProcessor
There was a bug in 0.0.7 where the bunkers do not defend. I refactored State Processor to allow bunkers to defend islands with no soldiers/squads. This however will require admin commands to set 1,0,0 on the blockchain.

I will changing 3 other gamestate processor sections though, so I won't push this to the community untill all of them are done.

Make sure to always to copy the sate processor before editing it, then rename the old state processor whatever version it is as its own class. This is so that the callbacks can always reference the different versions of state processors and process the game from scratch.