Changed rule so that players can only attack an island once.
The rules currently are, squads can only be up to 1 million health and attackers can have only 3 squads. But that kind of doesnt matter if you can just keep launching attacks until you win. So we changed the rule that you have one chance to launch an attack and win or lose, the attack island is removed from your list.

Co-Authored-By: Codie Petersen <codie-petersen@users.noreply.github.com>