Fixed Can't Remove Squads and Reinforcement Bugs
There were two bugs, one where you couldn't remove squads already present on the island and one where you couldn't reinforce islands with new squads.

The first was a simple case of a boolean switch in the SetNameAndType.
The Second was happening because the queued action was being initialized with all of the already defending squads.

It is fixed now.
----------------------------- IMPORTANTE ----------------------------
I also changed the callbacks to check for a certain height AND version of the state processor. Because I was testing on the new state processor on the real blockchain, it made differences in the old game state and the new game state, so I just added the height requirement.