Fixed bunker and squad visuals, also fixed gamestate bug.
There was a bug that was causing crashing on the gamestate processor. It is fixed now. But basically it was a index out of range error because defender in Conflict function is -1 if there are no defenders and i was checking index -1.

The other two bugs were simple visual bugs.