Fixed Island Management and hid advanced options.
There was a bug where any island that was lost through selling to the bank or war, would cause a crash if you were in island management.

There was also some confusion about how to login and what needed to be entered, so I just took that out, hid it, and put a simple connect button.