Updated GameStateProcessor 0.2.9
There was a bug where attackers were not capturing islands with nothing on it. It was happening because the value of "defendersLeft" was going negative and I was checking for defendersLeft == 0. I also changed "hasBunkers" up a bit because it was broken even though no bugs had been caused by it yet. BUT, there was bound to be a bug eventually. I was checking string against string, but the defense tiles could sometimes have blockers which would cause the string vs string to think there were bunkers even though there were none. So I made funciton in validity to check for bunkers character by character. No validity functions were altered only added.