Fixed Clean Battle Hud, Disappear Attack, Clean After Submit
There were several bugs concerning the battle hud. Revisiting the island exposed several variables that were not being reset properly. Additionally having a defense plan in the queued actions already in the list and trying to create an attack island was a missed state that I was not handling and it caused a bug where the main island would disappear. The player is now given a message from the pushNotification system to submit existing plans.

There are still bugs regarding the revisiting defense and attack islands. In the defense, there is a bug that LoadQueued Plans tries to load squads but the for loop is using existing defenders causing an out of range bug. And on the attack island there is a bug that is not loading attack plans correctly. I'll have to check that out, not sure what that is all about. But those are being fixed after this commit.