Message Fix
There was a message that I was lazy and hadn't fixed, ended up forgeting about it. But essentially, you can only have two bunkers per tile. When adding a third, it will fail and give an error message to the player in the notification. Exactly what it is supposed to do, but instead of saying there are two bunkers I was telling them to check resources.

So added a case to check how many bunkers. Then tell them if it is because of low resources or too many bunkers.