Updated GameStateProcessor
Fixed a remove squad problem. There was a problem when checking if "canRemove". I for some reason was checking also if the squad to remove was less than the length of the squads to remove array. It apprently made sense at the time, but is clearly wrong. that's fixed now.