GSP Fix and Notification Expansion.
There was a bug with the GSP where attacking a tile with only bunkers and the opfor winning, caused a out of index error because of -1 defender. I fixed it.

I am also expanding upon the Notification System. It is not complete yet. But I am heading to bed and I haven't pushed this GSP fix yet. The notifications will still work as before, I haven't changed the function over to the new one yet, but I am building a new function and the old function is still in use, so it works the same way still.