Fixed game state processor.
There was a bug in the game state processor, again dealing with the defender being a negative value and getting index error. There is also another bug that subtracts from defendersLeft even if you have not attacked an actual squad just a bunker, so i fixed that (I think haven't tested too much) by moving the decrementer into line 954 so its inside the check for squads.