Made a new server scene for server builds. Fixed notifications.
So we needed an easy way to pull stats and users from the database, so I made a scene specifically for server building. No graphics, only the GSP running, not even a splash screen. All that you need to do is click all of the scenes off in build settings and make sure only the server scene is on.

Also, I forgot to push, but there is some older fixes for the last release in the notification system. Converted the whole thing to a bunch of functions because the outputlog in unity was not matching up with my lines in my code, so I had to narrow it down to function because I wasn't able to recreate the bug myself.