Changed Transaction Data
Daniel suggested having the buyer submit the contract proposal so I had changed the code a bit, but I forgot to reflect that in the TransactionData class. The only thing that is needed now for that one is contract data.

Pretty sure at least. Might be somehting I add on later if I found out I need it. But pretty sure just the actual contract will have enough information.