Latest GUI Fix and Telemetry Updates
I haven't pushed in a while, so this has a couple of things in it.

There were some swapped GUI values in the pools so that is one important fix, I initially, had changed them to 0,1,2 as metal oil concrete, but then there were bugs and I realized it was wrong, changes everything back to oil metal concrete. But some how the GUI values are normal now. I have no clue why. So that is something to consider.

Also the telemetry has been changed a bit. I've added a couple of payload classes and some testing stuff. The class changes required some changing of other scripts related to telemetry so that is mostly what the changes in those classes are.