Atomic Transaction Code
Made all of the code for atomic transactions. It is not working but I think that is because the NameRawTransaction function returns a json object with "hex":value and I am actually trying to accept a string. It is returning null for some reason, and I think that is it. But all of the supporting functions have been made in the BitCoinLib project. I just need to test them out.