Restructured Namespaces and Contract Resolver Development
I broke up the huge namespaces into smaller files, they are still in the same namespaces but I saw the structure of BitCoinLib and it feels more organized even though it has more files. So I converted the namespaces to that style.

Also I have editted the ContractResolver and renamed it to TransactionResolver.  Renamed it because there is something else already named ContractResolver in regards to Json. So wanted to avoid naming confusion.

Mostly the Transaction Resolver is a payload packager and should check to see if there are any problems with the commander.