MarketScene And Scripts
I should have just pushed the scene then the scripts, but I didn't think about that. So there are some unfinished scripts right now for the market stuff.

I have a MarketClient which is basically a slimmed down version of the ClientInterface but only for market actions.

There are also Button Scripts for Menu Navigation.

There is a bunch of new GUI textures and they are used in the MarketScene.
Lastly the Market Scene has pretty much the final product of the Resource Market Menu in terms of visuals. The scripts for it still need to be built out.