Added Xaya Wallet Monitoring and Icon in the Market  Scene
Because the Chi market will require players to see how much chi they have available I have made an Icon that tracks the amount of chi  in their wallet. It's tracked in the ResourceMonitor script. The I added a function to Expose the CommsInterface's GetFunds() function in the MarketClient as well.