Reorganized Market Folder Made Chi Market Scripts
I came to realize that the ChiOrderItem will be nothing like the OrderItem. So, I had to rewrite some stuff and make a more generic base class. I haven't refactored the ResourceMarket, but I will/can after this ChiMarket is working. But for right now most of the generic stuff was mostly lift and shifts of MarketTrading stuff.

I changed some random variables and code here and there to make it look nicer in some places.

Additionally, I created a function that converts OrderPayload list from telemetry to a list of ChiOrderData Dictionary. Easier to find and manipulate orders.