Chi Order Updates
Touched a couple of things to get Chi Trading up and running. Tried to keep it as similar as possible to the Resource Trading scripts.

Many of the functions and process are the same. Got the formation gui working. Doesn't actually create an order though. I'll connect that stuff later.